==> 1_data_preprocessing.ipynb <==
import pandas as pd

# import data from Excel
data = pd.read_excel("C:\\Users\\Tim\Desktop\\Data Science\\Scriptie\\Tetris data\\dataset_for_data_mining.xlsx")

# print the data
print(data.head())


# Function to modify column names
def format_column_name(column_name):
    return column_name.lower().replace(" ", "_").replace("/", "_")

# Adjust the column names
data.columns = [format_column_name(col) for col in data.columns]

# Display the updated columns
print(data.columns)

# calculate median of Tetris experience column
median_experience = data["tetris_experience"].median()

# create a new column called "Experience level_binary" and set it to "Low" for values less than or equal to median, and "High" for values greater than median
data["tetris_experience_binary"] = [0 if x <= median_experience else 1 for x in data["tetris_experience"]]

# print the updated data
print(data["tetris_experience_binary"])

# calculate median of 'Videogames Hours a Week' column
median_experience = data["videogames_hours_a_week"].median()

# create a new column called "videogames_hours_a_week_binary" and set it to 0 for values less than or equal to median, and 1 for values greater than median
data["videogames_hours_a_week_binary"] = [0 if x <= median_experience else 1 for x in data["videogames_hours_a_week"]]

# print the updated data
print(data["tetris_experience_binary"])

# create a new column called "Tetris hours a week_binary" and set it to 0 for values that are zero, and 1 for values greater than zero.
data["tetris_hours_a_week_binary"] = [0 if x == 0 else 1 for x in data["tetris_hours_a_week"]]

# print the updated data
print(data["tetris_hours_a_week_binary"])


# Remove rows where the 'sex' column has a value other than 'Male' or 'Female'
data = data[data['sex'].isin(['Male', 'Female'])]

# Create a new binary column 'is_male' where 'Male' is 1 and 'Female' is 0
data['sex_num'] = data['sex'].map({'Male': 1, 'Female': 0})

# calculate median of 'Age' column
median_experience = data["age"].median()

# create a new column called "age_binary" and set it to 0 for values less than or equal to median, and 1 for values greater than median
data["age_binary"] = [0 if x <= median_experience else 1 for x in data["age"]]

# print the updated data
print(data["age_binary"])

from sklearn.cluster import KMeans

# K-means clustering of 'age'.
ages = data["age"].values.reshape(-1, 1)

# Choose the number of clusters
n_clusters = 2

# Perform k-means clustering
kmeans = KMeans(n_clusters=n_clusters, random_state=0).fit(ages)

# Get the cluster assignments for each age
cluster_assignments1 = kmeans.labels_

data['age_cluster'] = cluster_assignments1

# K-means clustering of 'videogames_hours_a_week'.
ages = data['videogames_hours_a_week'].values.reshape(-1, 1)

# Choose the number of clusters
n_clusters = 2

# Perform k-means clustering
kmeans = KMeans(n_clusters=n_clusters, random_state=0).fit(ages)

# Get the cluster assignments for each age
cluster_assignments = kmeans.labels_

data['videogames_hours_a_week_cluster'] = cluster_assignments

# Create 'total_keystrokes_averagepergame' column to divide the total number of keystrokes by the number of matched played.
data["total_keystrokes_averagepergame"] = data["total_keystrokes"] / data["matches_played"]

# Create 'total_down_averagepergame' column to divide the total down by the number of matched played.
data["total_down_averagepergame"] = data["total_down"] / data["matches_played"]

# Create 'total_left_averagepergame' column to divide the total left by the number of matched played.
data["total_left_averagepergame"] = data["total_left"] / data["matches_played"]

# Create 'total_right_averagepergame' column to divide the total right by the number of matched played.
data["total_right_averagepergame"] = data["total_right"] / data["matches_played"]

# Create 'total_left_right_averagepergame' column to divide the total left_right by the number of matched played.
data["total_left_right_averagepergame"] = data["total_left_right"] / data["matches_played"]

# Create 'total_clockwise_rotations_averagepergame' column to divide the total clockwise rotations by the number of matched played.
data["total_clockwise_rotations_averagepergame"] = data["total_clockwise_rotations"] / data["matches_played"]

# Create 'total_conterclockwise_rotations_averagepergame' column to divide the total conter clockwise rotations by the number of matched played.
data["total_conterclockwise_rotations_averagepergame"] = data["total_conterclockwise_rotations"] / data["matches_played"]

# Create 'total_rotations_averagepergame' column to divide the total rotations by the number of matched played.
data["total_rotations_averagepergame"] = data["total_rotations"] / data["matches_played"]

# Display the new columns
print(data.columns)

# save the dataframe as an Excel file
data.to_excel('output_preprocessed_data.xlsx', index=False)
