==> 4_split_stratified.ipynb <==
import pandas as pd

# import data from Excel
data = pd.read_excel("C:\\Users\\Tim\\Desktop\\Data Science\\Scriptie\\Tetris data\\code\\output_preprocessed_data.xlsx")

# print the columns
print(data.columns)


!pip install pandas numpy scikit-learn openpyxl
from sklearn.model_selection import train_test_split

# Define the columns to stratify by
stratify_columns = ['age_binary','sex_num', 'videogames_hours_a_week_binary']

# Calculate the combined stratification key
data['stratify_key'] = data[stratify_columns].apply(tuple, axis=1)

# Perform the first split: 80% train + validation, 20% test
train_val_data, test_data = train_test_split(data, test_size=0.2, stratify=data['stratify_key'], random_state=42)

# Perform the second split: 60% train, 20% validation (75% of the remaining 80%)
train_data, val_data = train_test_split(train_val_data, test_size=0.25, stratify=train_val_data['stratify_key'], random_state=42)

# Save the split datasets to separate Excel files
train_val_data.to_excel('train_val_data.xlsx', index=False)
train_data.to_excel('train_data.xlsx', index=False)
val_data.to_excel('val_data.xlsx', index=False)
test_data.to_excel('test_data.xlsx', index=False)
